requirements.txt
django>=5.0
djangorestframework >= 3.14.0
markdown
django-filter